% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLFD-package.R
\docType{package}
\name{PLFD-package}
\alias{PLFD}
\alias{PLFD-package}
\title{PLFD: A Portmanteau Local Feature Discrimination Approach to the High-Dimensional Matrix-Variate Data}
\source{
A Portmanteau Local Feature Discrimination Approach to the Classification 
with High-dimensional Matrix-variate Data.
}
\description{
Identify the local features that are significant to classification and employ them
    to form sub-discriminant functions, then aggregate the subfunctions for classification.
}
\details{
Identify the local features that are significant to classification and employ
them to form sub-discriminant functions, then aggregate the subfunctions for classification.
}
\keyword{internal}
