% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rFuns.R
\name{plfd}
\alias{plfd}
\title{Matrix-variate Linear Discriminant Analysis}
\usage{
plfd(
  x1,
  x2,
  rSize,
  cSize,
  blockList,
  blockMode,
  xtest,
  ytest,
  permNum = 50,
  alpha = 0
)
}
\arguments{
\item{x1}{Array of \eqn{r \times c \times n_1}, samples from group 1.}

\item{x2}{Array of \eqn{r \times c \times n_2}, samples from group 2.}

\item{rSize, cSize}{The common row size and column size of blocks. See details.}

\item{blockList}{List including the index of considered blocks. Optional if \code{rSize} and
\code{cSize} are provided. See details.}

\item{blockMode}{The mode how the nonzero positions in \eqn{M_1 - M_2} are detected. See details.}

\item{xtest}{(Optional)New samples to be predicted.}

\item{ytest}{(Optional)Vector with \eqn{1,2} entries corresponds to the labels of \code{xtest}, .}

\item{permNum}{The number of permutatation(default=50).}

\item{alpha}{The \code{alpha}-upper quantile of the permutated statistics of blocks under 
consideration is used(default=0).}
}
\value{
List, \itemize{
 \item \code{paras} List including the parameters of significant blocks.
 \item \code{y} Self-predicted results for training data. It is a matrix of \eqn{(n_1+n_2)\times 2}, 
         the first column is the scores and the second column is the predicted labels.
 \item \code{mcr} The self-predicted misclassification rate for training samples.
 \item \code{ytest.hat} The predicted result for \code{xtest} if it is provided. It is a
       matrix where the first column is scores and the second column is predicted group.
 \item \code{mcr.test} The misclassification rate for \code{xtest} if \code{ytest} is provided.
}
}
\description{
Matrix linear discriminant analysis.
}
\details{
There are two manners to specify the submatrix-variates under consideration. In the case that 
the matrix-variate is splitted into non-overlapped blocks that share the common row size and
column size, these sizes can be specified by \code{rSize} and \code{cSize}. Otherwise, the 
submatrix-variates can be flexibly specified by \code{blockList}, which should be a list that each 
component includes \code{rIdx} and \code{cIdx} corresponding to the rows index and columns 
index of a sumatrix-variate. See examples.

Three modes are provided to \code{blockMode}, \code{dense, fd(forward), bd(backward)}. 
If \code{blockMode=='dense'}, the nonzero positions are assumed as all positions. If 
\code{blockMode=='fd'} or \code{'bd'}, the nonzero positions are detected by forward/backward 
sequential likelihood method, in which BIC serves as the stoppting rule.
}
\examples{
r0 <- 20
c0 <- 25
n1 <- n2 <- 100
ntest <- 50
x1 <- array(rnorm(r0*c0*n1, mean=1), dim=c(r0, c0, n1))
x2 <- array(rnorm(r0*c0*n2, mean=2), dim=c(r0, c0, n2))
xtest <- array(rnorm(r0*c0*ntest, mean=2), dim=c(r0, c0, ntest))
ytest <- rep(2, ntest)

## Uniformly splitting
(result <- plfd(x1, x2, rSize=5, cSize=5, blockMode='dense', xtest=xtest, ytest=ytest))

## Specify the blocks
blockList <- list(list(rIdx=1:5, cIdx=1:5), 
                  list(rIdx=6:10, cIdx=1:5), 
                  list(rIdx=3:9, cIdx=2:8))
(result <- plfd(x1, x2, blockList=blockList, blockMode='dense', xtest=xtest, ytest=ytest))
}
