% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-plfd.R
\name{predict.plfd}
\alias{predict.plfd}
\title{Predict Method for \code{plfd}}
\usage{
\method{predict}{plfd}(object, x, y, ...)
}
\arguments{
\item{object}{\code{plfd} object.}

\item{x}{Array, matrix-variate data to be predicted.}

\item{y}{Vector (optional), Labels of \code{x} with value \code{1} or \code{2}.}

\item{...}{Ignored currently.}
}
\value{
\code{list(W, y.hat, mcr)}, \itemize{
\item \code{W}: discriminant scores;
\item \code{y.hat}: predicted labels;
\item \code{mcr}: misclassification rate if parameter \code{y} is available.
}
\description{
Predict Method for \code{plfd}
}
