% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print-plfd.R
\name{print.plfd}
\alias{print.plfd}
\title{Print Method for \code{plfd}}
\usage{
\method{print}{plfd}(x, ...)
}
\arguments{
\item{x}{\code{plfd} object.}

\item{...}{Ignored currently.}
}
\description{
Print Method for \code{plfd}
}
