% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLFD-package.R
\docType{package}
\name{PLFD-package}
\alias{PLFD}
\alias{PLFD-package}
\title{PLFD: Portmanteau Local Feature Discrimination for Matrix-Variate Data}
\description{
The portmanteau local feature discriminant approach first identifies the local discriminant features and their differential structures, then constructs the discriminant rule by pooling the identified local features together. This method is applicable to high-dimensional matrix-variate data. See the paper by Xu, Luo and Chen (2021, \doi{10.1007/s13171-021-00255-2}).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/paradoxical-rhapsody/PLFD}
  \item Report bugs at \url{https://github.com/paradoxical-rhapsody/PLFD/issues}
}

}
\author{
\strong{Maintainer}: Zengchao Xu \email{zengc.xu@aliyun.com}

Authors:
\itemize{
  \item Shan Luo
  \item Zehua Chen
}

}
\keyword{internal}
