% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLFD-package.R
\docType{package}
\name{PLFD-package}
\alias{PLFD}
\alias{PLFD-package}
\title{PLFD: A Portmanteau Local Feature Discrimination Approach to the High-Dimensional Matrix-Variate Data}
\source{
A Portmanteau Local Feature Discrimination Approach to the Classification
with High-dimensional Matrix-variate Data.
}
\description{
It is an implementation of the portmanteau-local-feature discrimination. See the source for details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitee.com/xu-zc/PLFD}
  \item Report bugs at \url{https://gitee.com/xu-zc/PLFD/issues}
}

}
\keyword{internal}
