% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PLFD-package.R
\docType{package}
\name{PLFD-package}
\alias{PLFD}
\alias{PLFD-package}
\title{PLFD: Portmanteau Local Feature Discrimination for Matrix-Variate Data}
\description{
The PLFD in first identifies the local discriminant features and their 
    differential structures, then constructs the discriminant rule by pooling 
    the identified local features together. This method is applicable to 
    high-dimensional matrix-variate data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gitee.com/xu-zc/PLFD}
  \item Report bugs at \url{https://gitee.com/xu-zc/PLFD/issues}
}

}
\author{
\strong{Maintainer}: Zengchao Xu \email{zengc.xu@aliyun.com}

Authors:
\itemize{
  \item Shan Luo
  \item Zehua Chen
}

}
\keyword{internal}
