% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plfd.R
\name{plfd}
\alias{plfd}
\title{Matrix-variate Linear Discriminant Analysis}
\usage{
plfd(
  x1,
  x2,
  r0,
  c0,
  blockList,
  blockMode = "dense",
  xtest,
  ytest,
  permNum = 50,
  alpha = 0
)
}
\arguments{
\item{x1}{Array of \eqn{r \times c \times n_1}, samples from group 1.}

\item{x2}{Array of \eqn{r \times c \times n_2}, samples from group 2.}

\item{r0, c0}{The common row size and column size of blocks. See details.}

\item{blockList}{List including the index of considered blocks. Optional if \code{r0} and
\code{c0} are provided. See details.}

\item{blockMode}{How the nonzero positions of \code{M1-M2} are detected.
Three modes are provided: \code{"dense"}(default), \code{"fd"/"forward"} and \code{"bd"/"backward"}.
The \code{blockMode=="dense"} supposes that \code{M1 - M2} has no zero entry.
If \code{blockMode=="fd"} or \code{"bd"}, the nonzero positions are detected by
forward/backward step, where \emph{BIC} serves as the stopping rule.}

\item{xtest}{(Optional) New samples to be predicted.}

\item{ytest}{(Optional) Vector with \eqn{1,2} entries corresponds to the labels of \code{xtest}.}

\item{permNum}{The number of permutation (default=50).}

\item{alpha}{The upper alpha` quantile of the permutation statistics (default=0).}
}
\value{
List, \itemize{
\item \code{paras} List including the parameters of significant blocks.
\item \code{y} Self-predicted results for training data. It is a matrix of \eqn{(n_1+n_2)\times 2},
the first column is the scores and the second column is the predicted labels.
\item \code{mcr} The self-predicted misclassification rate for training samples.
\item \code{ytest.hat} The predicted result for \code{xtest} if it is provided. It is a
matrix where the first column is scores and the second column is predicted group.
\item \code{mcr.test} The misclassification rate for \code{xtest} if \code{ytest} is provided.
}
}
\description{
Matrix linear discriminant analysis.
}
\details{
There are two manners to specify the blocks under consideration. In the case that
the matrix-variate is partition into non-overlapping blocks that share the common row size and
column size, these sizes can be specified by \code{r0} and \code{c0}. Otherwise, the
blocks can be flexibly specified by \code{blockList}, which should be a list that each
component includes \code{rIdx} and \code{cIdx} corresponding to the rows index and columns
index of a submatrix-variate. See examples.
}
\examples{
## Simulate the data of small dimension and sample size for saving time
set.seed(2020)
rDim <- 20
cDim <- 20
n1 <- n2 <- 50
ntest <- 50
x1 <- array(rnorm(rDim*cDim*n1, mean=0.0), dim=c(rDim, cDim, n1))
x2 <- array(rnorm(rDim*cDim*n2, mean=1.0), dim=c(rDim, cDim, n2))
xtest <- array(rnorm(rDim*cDim*ntest, mean=1.0), dim=c(rDim, cDim, ntest))
ytest <- rep(2, ntest)

## Uniform partition
(result <- plfd(x1, x2, r0=5, c0=5, blockMode='dense', xtest=xtest, ytest=ytest))

## Pre-specify feature blocks
blockList <- list(list(rIdx=1:5, cIdx=1:5), 
                  list(rIdx=6:10, cIdx=1:5), 
                  list(rIdx=3:9, cIdx=2:8))
(plfd.model <- plfd(x1, x2, blockList=blockList, blockMode='dense', xtest=xtest, ytest=ytest))

## print
print(plfd.model)

## Predict
predict(plfd.model, xtest)

}
