% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-plfd.R
\name{predict.plfd}
\alias{predict.plfd}
\title{Predict Method for \code{plfd}}
\usage{
\method{predict}{plfd}(object, x, y, ...)
}
\arguments{
\item{object}{\code{plfd} object.}

\item{x}{The samples to be predicted.}

\item{y}{Vector (optional). Labels of x with value \code{1} or \code{2}.}

\item{...}{Ignored currently.}
}
\value{
\code{list(W, y.hat, mcr)}, wherein \code{W} refers to the discriminant
scores, \code{y.hat} refers to the predicted labels and \code{mcr} is the misclassification
rate when \code{y} is available.
}
\description{
Predict Method for \code{plfd}
}
