% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-plfd.R
\name{predict.plfd}
\alias{predict.plfd}
\title{Predict Method for \code{plfd}}
\usage{
\method{predict}{plfd}(object, x, ...)
}
\arguments{
\item{object}{\code{plfd} object.}

\item{x}{The samples to be predicted.}

\item{...}{Ignored currently.}
}
\value{
Matrix with two columns, including the scores and predicted groups.
}
\description{
Predict Method for \code{plfd}
}
