% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rFuns.R
\name{predict.plfd}
\alias{predict.plfd}
\title{Predict Method for \code{plfd}}
\usage{
\method{predict}{plfd}(object, x, ...)
}
\arguments{
\item{object}{\code{plfd} object.}

\item{x}{The samples to be predicted.}

\item{...}{Ignored currently.}
}
\value{
Matrix with two columns, including the scores and predicted groups.
}
\description{
Predict Method for \code{plfd}
}
\examples{
r0 <- 20
c0 <- 25
n1 <- n2 <- 100
x1 <- array(rnorm(r0*c0*n1, mean=1), dim=c(r0, c0, n1))
x2 <- array(rnorm(r0*c0*n2, mean=2), dim=c(r0, c0, n2))
(plfd.model <- plfd(x1, x2, rSize=5, cSize=5, blockMode='dense'))

ntest <- 50
xtest <- array(rnorm(r0*c0*ntest, mean=2), dim=c(r0, c0, ntest))
predict(plfd.model, xtest)
}
